bb-imager-gui: More refactoring
- Do not put single use stuff in helpers
