+++ bb-imager-gui/src/helpers.rs
@@ -1,7 +1,6 @@
-use std::{
-    borrow::Cow, collections::HashSet, fmt::Display, io::Read, path::PathBuf, sync::LazyLock,
-};
+use std::{borrow::Cow, fmt::Display, path::PathBuf, sync::LazyLock};
 
+use crate::BBImagerMessage;
 use bb_config::config::{self, OsListItem};
 use bb_flasher::{BBFlasher, BBFlasherTarget, DownloadFlashingStatus};
 use futures::StreamExt;
@@ -9,10 +8,6 @@ use iced::{
     futures,
     widget::{self, text},
 };
-use serde::{Deserialize, Serialize};
-use tokio::io::AsyncWriteExt;
-
-use crate::{BBImagerMessage, constants};
 
 #[derive(Clone, Debug, Default, PartialEq)]
 pub(crate) struct ProgressBarState {
@@ -322,222 +317,7 @@ impl std::fmt::Display for BoardImage {
 
 pub(crate) fn system_timezone() -> Option<&'static String> {
     static SYSTEM_TIMEZONE: LazyLock<Option<String>> = LazyLock::new(localzone::get_local_zone);
-
     (*SYSTEM_TIMEZONE).as_ref()
-}
-
-/// Configuration for GUI that should be presisted
-#[derive(Default, Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct GuiConfiguration {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    sd_customization: Option<SdCustomization>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    bcf_customization: Option<BcfCustomization>,
-    #[cfg(feature = "pb2_mspm0")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pb2_mspm0_customization: Option<Pb2Mspm0Customization>,
-}
-
-impl GuiConfiguration {
-    pub(crate) fn load() -> std::io::Result<Self> {
-        let mut data = Vec::with_capacity(512);
-        let config_p = Self::config_path().unwrap();
-
-        let mut config = std::fs::File::open(config_p)?;
-        config.read_to_end(&mut data)?;
-
-        Ok(serde_json::from_slice(&data).unwrap())
-    }
-
-    pub(crate) async fn save(&self) -> std::io::Result<()> {
-        let data = serde_json::to_string_pretty(self).unwrap();
-        let config_p = Self::config_path().unwrap();
-
-        tracing::info!("Configuration Path: {:?}", config_p);
-        tokio::fs::create_dir_all(config_p.parent().unwrap()).await?;
-
-        let mut config = tokio::fs::OpenOptions::new()
-            .write(true)
-            .create(true)
-            .truncate(true)
-            .open(config_p)
-            .await?;
-
-        config.write_all(data.as_bytes()).await?;
-
-        Ok(())
-    }
-
-    fn config_path() -> Option<PathBuf> {
-        let dirs = directories::ProjectDirs::from(
-            constants::PACKAGE_QUALIFIER.0,
-            constants::PACKAGE_QUALIFIER.1,
-            constants::PACKAGE_QUALIFIER.2,
-        )?;
-
-        Some(dirs.config_local_dir().join("config.json").to_owned())
-    }
-
-    pub(crate) const fn sd_customization(&self) -> Option<&SdCustomization> {
-        self.sd_customization.as_ref()
-    }
-
-    pub(crate) const fn bcf_customization(&self) -> Option<&BcfCustomization> {
-        self.bcf_customization.as_ref()
-    }
-
-    #[cfg(feature = "pb2_mspm0")]
-    pub(crate) const fn pb2_mspm0_customization(&self) -> Option<&Pb2Mspm0Customization> {
-        self.pb2_mspm0_customization.as_ref()
-    }
-
-    pub(crate) fn update_sd_customization(&mut self, t: SdCustomization) {
-        self.sd_customization = Some(t);
-    }
-
-    pub(crate) fn update_bcf_customization(&mut self, t: BcfCustomization) {
-        self.bcf_customization = Some(t)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub(crate) struct SdCustomization {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) hostname: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) timezone: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) keymap: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) user: Option<SdCustomizationUser>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) wifi: Option<SdCustomizationWifi>,
-}
-
-impl SdCustomization {
-    pub(crate) fn update_hostname(mut self, t: Option<String>) -> Self {
-        self.hostname = t;
-        self
-    }
-
-    pub(crate) fn update_timezone(mut self, t: Option<String>) -> Self {
-        self.timezone = t;
-        self
-    }
-
-    pub(crate) fn update_keymap(mut self, t: Option<String>) -> Self {
-        self.keymap = t;
-        self
-    }
-
-    pub(crate) fn update_user(mut self, t: Option<SdCustomizationUser>) -> Self {
-        self.user = t;
-        self
-    }
-
-    pub(crate) fn update_wifi(mut self, t: Option<SdCustomizationWifi>) -> Self {
-        self.wifi = t;
-        self
-    }
-}
-
-impl From<SdCustomization> for bb_flasher::sd::FlashingSdLinuxConfig {
-    fn from(value: SdCustomization) -> Self {
-        Self::new(
-            value.hostname,
-            value.timezone,
-            value.keymap,
-            value.user.map(|x| (x.username, x.password)),
-            value.wifi.map(|x| (x.ssid, x.password)),
-        )
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct SdCustomizationUser {
-    pub(crate) username: String,
-    pub(crate) password: String,
-}
-
-impl SdCustomizationUser {
-    pub(crate) const fn new(username: String, password: String) -> Self {
-        Self { username, password }
-    }
-
-    pub(crate) fn update_username(mut self, t: String) -> Self {
-        self.username = t;
-        self
-    }
-
-    pub(crate) fn update_password(mut self, t: String) -> Self {
-        self.password = t;
-        self
-    }
-}
-
-impl Default for SdCustomizationUser {
-    fn default() -> Self {
-        Self::new(whoami::username(), String::new())
-    }
-}
-
-#[derive(Default, Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct SdCustomizationWifi {
-    pub(crate) ssid: String,
-    pub(crate) password: String,
-}
-
-impl SdCustomizationWifi {
-    pub(crate) fn update_ssid(mut self, t: String) -> Self {
-        self.ssid = t;
-        self
-    }
-
-    pub(crate) fn update_password(mut self, t: String) -> Self {
-        self.password = t;
-        self
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct BcfCustomization {
-    pub(crate) verify: bool,
-}
-
-impl BcfCustomization {
-    pub(crate) fn update_verify(mut self, t: bool) -> Self {
-        self.verify = t;
-        self
-    }
-}
-
-impl Default for BcfCustomization {
-    fn default() -> Self {
-        Self { verify: true }
-    }
-}
-
-#[cfg(feature = "pb2_mspm0")]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct Pb2Mspm0Customization {
-    pub(crate) persist_eeprom: bool,
-}
-
-#[cfg(feature = "pb2_mspm0")]
-impl Pb2Mspm0Customization {
-    pub(crate) fn update_persist_eeprom(mut self, t: bool) -> Self {
-        self.persist_eeprom = t;
-        self
-    }
-}
-
-#[cfg(feature = "pb2_mspm0")]
-impl Default for Pb2Mspm0Customization {
-    fn default() -> Self {
-        Self {
-            persist_eeprom: true,
-        }
-    }
 }
 
 #[derive(Debug, Clone)]
@@ -722,7 +502,7 @@ impl Destination {
     }
 }
 
-pub(crate) async fn destinations(flasher: config::Flasher) -> HashSet<Destination> {
+pub(crate) async fn destinations(flasher: config::Flasher) -> Vec<Destination> {
     match flasher {
         config::Flasher::SdCard => bb_flasher::sd::Target::destinations()
             .await
@@ -795,18 +575,18 @@ const fn flasher_supported(flasher: config::Flasher) -> bool {
 #[derive(Clone, Debug)]
 pub(crate) enum FlashingCustomization {
     LinuxSdFormat,
-    LinuxSd(SdCustomization),
-    Bcf(BcfCustomization),
+    LinuxSd(crate::persistance::SdCustomization),
+    Bcf(crate::persistance::BcfCustomization),
     Msp430,
     #[cfg(feature = "pb2_mspm0")]
-    Pb2Mspm0(Pb2Mspm0Customization),
+    Pb2Mspm0(crate::persistance::Pb2Mspm0Customization),
 }
 
 impl FlashingCustomization {
     pub(crate) fn new(
         flasher: config::Flasher,
         img: &BoardImage,
-        app_config: &GuiConfiguration,
+        app_config: &crate::persistance::GuiConfiguration,
     ) -> Self {
         match flasher {
             config::Flasher::SdCard if img.is_sd_format() => Self::LinuxSdFormat,
