+++ bb-imager-gui/src/ui/home.rs
@@ -5,7 +5,7 @@ use iced::{
 
 use crate::{BBImagerMessage, constants, helpers, pages::ImageSelectionState};
 
-use super::helpers::{home_btn_svg, home_btn_text};
+use super::helpers::home_btn_text;
 use crate::pages::Screen;
 
 pub(crate) fn view<'a>(
@@ -133,4 +133,48 @@ pub(crate) fn view<'a>(
         .into()
     })
     .into()
+}
+
+pub(crate) fn home_btn_svg<'a>(
+    icon: &'static [u8],
+    active: bool,
+) -> widget::Button<'a, BBImagerMessage> {
+    const ICON_SIZE: u16 = 32;
+    const PADDING: u16 = 4;
+    const RADIUS: u16 = (ICON_SIZE + PADDING * 2) / 2;
+
+    fn svg_style(active: bool) -> widget::svg::Style {
+        if active {
+            Default::default()
+        } else {
+            widget::svg::Style {
+                color: Some(iced::Color::BLACK.scale_alpha(0.5)),
+            }
+        }
+    }
+
+    fn btn_style(active: bool) -> widget::button::Style {
+        if active {
+            widget::button::Style {
+                background: Some(iced::Color::WHITE.into()),
+                border: iced::border::rounded(RADIUS),
+                ..Default::default()
+            }
+        } else {
+            widget::button::Style {
+                background: Some(iced::Color::BLACK.scale_alpha(0.5).into()),
+                border: iced::border::rounded(RADIUS),
+                ..Default::default()
+            }
+        }
+    }
+
+    widget::button(
+        widget::svg(widget::svg::Handle::from_memory(icon))
+            .style(move |_, _| svg_style(active))
+            .width(ICON_SIZE)
+            .height(ICON_SIZE),
+    )
+    .style(move |_, _| btn_style(active))
+    .padding(PADDING)
 }
